Add fleet coordinate management and resource handling; implement methods for adding and withdrawing ammo and provision, and create new DTOs for fleet movement and attack results

+++ src/main/java/com/hush0k/pirateTeam/fleet/dto/response/FleetStatsDto.java
@@ -6,7 +6,7 @@ public record FleetStatsDto(
         int boardingPower,
         int manoeuvrability,
         int combatStability,
-        int lootMultiplier
+        float lootMultiplier
 
 ) {}
